+++ CHANGELOG.md
@@ -1,5 +1,237 @@
-# Changelog
+# CHANGELOG
 
-## [0.1] - 13 / 08 / 2024
+## [0.31] - 23 / 09 / 2024 - Test 
 
-## Agregado
+### Added
+
+- Test knight
+
+## [0.30] - 22 / 09 / 2024 - Knight 
+
+### Added
+
+- Knight, movimientos caballo
+
+## [0.29] - 21 / 09 / 2024 - Tests 
+
+### Added
+
+- Test Queen
+
+## [0.28] - 20 / 09 / 2024 - Tests arreglos 
+
+### Fixed
+
+- Test Torres
+- Test Bishop
+
+## [0.27] - 19 / 09 / 2024 - Codigo repetido
+
+### Changed
+
+- Pieces
+- Rook 
+- Bishop
+- Queen
+- King
+
+## [0.26] - 18 / 09 / 2024 - Testeos varios
+
+### Fixed
+
+- test cli
+
+### Added
+
+- test excepciones
+- test chess
+
+## [0.25] - 17 / 09 / 2024 - Movimientos piezas
+
+### Added
+ 
+- Movimientos clase Queen, King, Knight, Bishop
+
+## [0.24] - 16 / 09 / 2024 - Pieces
+
+### Fixed
+ 
+- clase pieces
+
+### Added
+
+- Test pieces
+
+## [0.23] - 15 / 09 / 2024 - Pawns
+
+### Fixed
+ 
+- Clase peon
+- Test peon movimientos
+
+## [0.22] - 12 / 09 / 2024 - Movimientos horizontales rook
+
+### Fixed
+ 
+- Test torres movimiento vertical en descenso
+
+### Added
+
+- Movimientos horizontales clase rook
+
+## [0.21] - 08 / 09 / 2024 - Tests
+
+### Added
+
+- Test funcion "capturar pieza"
+
+## [0.20] - 07 / 09 / 2024 - Arreglos
+
+### Changed
+
+- Arreglos en clase bishop, king, knight, pawn, rook
+- Arreglos en clase board, pieces, excepcions
+- Arreglos test caballo, test cli, test board, test torres
+
+## [0.19] - 06 / 09 / 2024 - Simbolos piezas
+
+### Changed
+
+- clase chess excepciones
+
+### Added
+
+- Simbolos piezas
+
+## [0.18] - 05 / 09 / 2024 - Cambios
+
+### Changed
+
+- Funcion "get piece" del board
+- Clase rook posiciones posibles
+- Excepciones 
+- Test board
+
+### Added
+
+- Funcion "move" del board
+- Clase pawn piezas
+- Funcion "color" clase chess
+
+## [0.17] - 04 / 09 / 2024 - Knight Test
+
+### Added
+
+- Test clase caballo- movimientos validos e invalidos
+
+## [0.16] - 02 / 09 / 2024 - Knight 
+
+### Added
+
+- Primeros movimientos clase caballo
+
+## [0.15] - 01 / 09 / 2024 - Arreglos board 
+
+### Changed
+
+- Test movimientos en clase board
+
+## [0.14] - 31 / 08 / 2024 - Cambios 
+
+### Changed
+
+- Test movimientos invalidos en board
+- Test movimiento piezas (rook) en board
+
+## [0.13] - 30 / 08 / 2024 - Cambios 
+
+### Changed
+
+- Modificaciones movimientos invalidos y validacion de coords (chess)
+- Excepciones
+
+## [0.12] - 29 / 08 / 2024 - Cambios 
+
+### Changed
+
+- Cambios en clase piezas
+- Movimientos clase torre
+
+### Added
+
+- Funciones en clase board
+
+## [0.11] - 25 / 08 / 2024 - Alfil
+
+### Added
+
+- Movimientos clase alfil
+- Test Alfil
+- Documentacion clase torre y test torres
+
+## [0.10] - 24 / 08 / 2024 - Tablero
+
+### Added
+
+- Test torres movimientos
+- Test Board movimientos
+- Posiciones en el tablero de todas las fichas
+
+## [0.9] - 23 / 08 / 2024 - Tests Interfaz
+
+### Added
+
+- Test cli
+
+## [0.8] - 22 / 08 / 2024 - Modificaciones
+
+### Fixed
+
+- Movimientos completos de la torre
+
+## [0.7] - 21 / 08 / 2024 - Excepciones
+
+### Changed
+
+- Excepciones agrupadas
+- Excepciones estandar y genericas 
+
+## [0.6] - 20 / 08 / 2024 - Modificaciones
+
+### Added
+
+- testeos movimientos del usuario
+- clase excepciones
+
+## [0.5] - 19 / 08 / 2024 - Turnos
+
+### Added
+
+- Turnos y excepciones
+- Verificar que la pieza existe en el tablero
+
+## [0.4] - 18 / 08 / 2024 - Torre
+
+### Added
+
+- Movimientos validos e invalidos
+- Primera prueba de test Torres
+
+## [0.3] - 17 / 08 / 2024 - Separacion clases
+
+### Added
+
+- Clases de las piezas en archivos separados
+
+## [0.2] - 16 / 08 / 2024 - Posiciones
+
+### Added
+
+- Agregado de posiciones en el board (Knight, Queen, King, Pawns)
+
+
+## [0.1] - 15 / 08 / 2024 - Primeros cambios
+
+### Added
+
+- Agregado posiciones de alfiles en el board